requirements.txt
pygame
pygame_gui
pygame_menu
chess
berserk-downstream
requests
requests_oauthlib